In case of bad_object, Mnesia needs to be restarted to recover
This makes a call into the kred/kdb system in order to
be able to restart Mnesia when a bad_object error occurs.

This creates an unfortunate dependency on Kred in OTP.
In particular, running OTP's stdlib test suite doesn't work
since the test suite triggers this condition, and then crashes
due to the undefined function call. This is handled by checking
if the callee is exported before calling it.